Find Shortest Duration Trial

Find the trial with the shortest duration in calendar time with a specified power and probability that RMST difference will be estimable with the Kaplan-Meier estimator. Within all trials of that minimum duration, the function finds the one with the smallest sample size. This function is based on a one-sided test of the superiority of the treatment arm.

shortest_duration(survdefC, survdefT, tau, power, accrual_rate,
  pKME = 0.95, alpha = 0.025, altdesign = F, multiplier = 1.1)

Arguments

survdefC
the survival distribution of the control group, as a list in the form output by survdef.

survdefT
the survival distribution of the treatment group, as a list in the form output by survdef.

tau
restriction time for the RMST.

power
the desired power.

accrual_rate
the planned accrual rate, per unit of time.

pKME
The desired probability that the RMST difference will be estimable using the Kaplan-Meier estimator. Default is 0.95.

alpha
one-sided type I error level. Default is 0.025.

altdesign
if T, the function will look for an alternative design that is not the shortest in duration, but has duration equal to some multiple of the shortest possible duration. Default is F. The sample size of the shortest duration trial can be much larger than a slightly longer trial, so we recommend considering an alternative design slightly longer than the shortest trial in addition to the shortest trial.

multiplier
the factor by which the duration of the shortest possible trial is multiplied to acquire the duration of the alternative trial design. Default is 1.1, meaning a trial that is 10 shortest possible trial. This argument is ignored if altdesign=F.

Value

a list with components:

n

total number of patients.

k1

length of the accrual period. We assume subjects will accrue uniformly over the interval (0, k1) and then be followed until trial time k1+k2.

k2

length of the follow-up period.

duration

trial duration in calendar time, k1+k2.

powerRMSTToverC

the asymptotic power of the RMST test for the superiority of treatment over control.

powerLRToverC

the asymptotic power of the log-rank test using all available follow-up for the superiority of treatment over control.

powerLRtauToverC

the asymptotic power of the log-rank test using follow-up to time tau for the superiority of treatment over control.

pKME

the probability that you will be able to estimate RMST difference at time tau using the standard Kaplan-Meier estimator.

Examples

con<-survdef(times = 3, surv = 0.5)
trt<-survdef(haz = 0.67*con$h(1))
shortest_duration(con, trt, 3, .8, 552/4)

#> $n
#> [1] 668
#> 
#> $k1
#> [1] 4.84058
#> 
#> $k2
#> [1] 0
#> 
#> $duration
#> [1] 4.84058
#> 
#> $powerRMSTToverC
#> [1] 0.8006054
#> 
#> $powerLRToverC
#> [1] 0.8586702
#> 
#> $powerLRtauToverC
#> [1] 0.8219996
#> 
#> $pMKE
#> [1] 1
#>