Create a Survdef Object Based on a Hazard Ratio

Creates a new object which stores user-specified survival distribution information in the format needed for the main function, RMSTpow. survdefHR is used when the user wishes to specify a survival distribution that is defined by its relationship to another distribution via a constant hazard ratio.

survdefHR(survdefC, HR)

Arguments

survdefC
the survival distribution for the reference/control group, as a list in the form output by survdef.

HR
the hazard ratio defining the relationship between the two distributions.

Value

a list with components:

S

a vectorized function that takes time as input and returns the survival probability at that time

h

a vectorized function that takes time as input and returns the hazard at that time

Examples

con<-survdef(times = 3, surv = 0.5); survdefHR(con, 0.5)

#> $S
#> function (x) 
#> sapply(x, function(t) {
#>     if (t == 0) 
#>         1
#>     else exp(-integrate(trthaz, lower = 0, upper = t)$value)
#> })
#> <environment: 0x7fb6674de238>
#> 
#> $h
#> function (t) 
#> survdefC$h(t) * HR
#> <environment: 0x7fb6674de238>
#>