Sample Size and Power for the Test of the Difference in Restricted Mean Survival Time

Determine the asmymptotic power of the test of RMST under a given trial design, or calculate the samples size needed to achieve a desired power.

RMSTpow(survdefC, survdefT, k1, k2, tau, n = NA, power = NA,
  plot = F, sim = F, M = 1000, method = "tau_star",
  alpha = 0.025)

Arguments

survdefC
the survival distribution of the control group, as a list in the form output by survdef.

survdefT
the survival distribution of the treatment group, as a list in the form output by survdef.

k1
length of the accrual period. We assume subjects will accrue uniformly over the interval (0, k1) and then be followed until trial time k1+k2.

k2
length of the follow-up period.

tau
restriction time for the RMST.

n
total sample size for both groups. 1:1 randomization is assumed. Either n or power can be specified, and the other value will be calculated.

power
the desired power.

plot
if T, plots of power versus each design parameter will be produced. Default is F.

sim
if T, simulations will be conducted and empirical power and other summary statistics will be provided. Default is F. The hypothesis tests are carried out based on the normal approximation with the variance estimated according to the Greenwood plug-in/infinitesimal jackknife method. Specifying situations where survival doesn't go to zero in a reasonable amount of time (trial length times 1000) will lead to problems if the sim=T option is used.

M
number of simualations. Default is 1000.

method
modification to be used in simulations if the Kaplan-Meier estimator is not defined at time tau in either group. Default is 'tau_star', which changes the restriction time to the last censoring time, if the last observation is censored at a time earlier than tau. Other possible values are 'gill', 'efron', 'tau_star', 'risk1', 'risk2', and 'risk5'. The riskX' options indicate estimating RMST difference at the latest time at which at least X people are at risk in each group, irrespective of the value of tau.

alpha
one-sided type I error level. Default is 0.025.

Value

a list with components

n

the user specified n, or if n was left blank, the n needed to achieve the user-specified power.

powerRMST

the user specified power, or if power was left blank, the asymptotic power of the RMST test.

powerLR

the asymptotic power of the log-rank test.

pKME

the probability that you will be able to estimate RMST difference at time tau using the standard Kaplan-Meier estimator.

plotvals

a list used for generating plots, returned if plot = T, with components:

n_vec a vector of sample sizes near the selected sample size n_pow power of the RMST based test for the values in n_vec, holding k1, k2 and tau constant k1_vec a vector of accrual periods near the specified k1 k1_pow power of the RMST based test for the values in k1_vec, holding n, tau and total trial length constant k2_vec a vector of follow-up periods near the specified k2 k2_pow power of the RMST based test for the values in k2_vec, holding n, k1 and tau constant tau_vec a vector of restriction times near the specified tau tau_pow power of the RMST based test for the values in tau_vec, holding n, k1 and k2 constant
simout

a list returned if sim = T, with components:

emppowRMST empirical power of the RMST test. emppowLR empirical power of the log-rank test. emppKME proportion of simulations where the standard KM estimator was used. meandiff mean estimated difference in RMST across the simulated datasets. SDdiff standard deviation of the estimated difference in RMST across the simulated datasets. meantrunc mean truncation time used in the simulated datasets (may be smaller than tau if method = 'tau_star' or 'riskX' options are used). SDtrunc standard deviation of the truncation time used in the simulated datasets.
Examples

con<-survdef(times = 3, surv = 0.5)
trt<-survdef(haz = 0.67*con$h(1))
RMSTpow(con, trt, k1 = 0, k2 = 3, tau = 3, power = 0.8)

#> $n
#> [1] 552
#> 
#> $powerRMST
#> [1] 0.8008182
#> 
#> $powerLR
#> [1] 0.870807
#> 
#> $pKME
#> [1] 1
#> 

RMSTpow(con, trt, k1 = 0, k2 = 3, tau = 3, n = 552)

#> $n
#> [1] 552
#> 
#> $powerRMST
#> [1] 0.8008182
#> 
#> $powerLR
#> [1] 0.870807
#> 
#> $pKME
#> [1] 1
#>